fix: move ArrayViewArgHandler in a common placewq
